# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Transfer Sederhana",
        "description": "Dokumentasi API Transfer Sederhana, dibuat oleh Aditya Nanda Utama untuk keperluan seleksi BackEnd Programmer di bosCOD",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "aditya.nanda0030@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login Auth",
                "description": "Login Auth",
                "operationId": "loginAuth",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "aditya.nanda0030@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register Auth",
                "description": "Register Auth",
                "operationId": "registerAuth",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "aditya.nanda0030@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Aditya Nanda"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Me Auth",
                "description": "Me Auth",
                "operationId": "meAuth",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/update-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update Token Auth",
                "description": "Update Token Auth",
                "operationId": "updateTokenAuth",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "dhaiuhqijdqjdowd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/bank/": {
            "get": {
                "tags": [
                    "Bank"
                ],
                "summary": "All Bank",
                "description": "All Bank",
                "operationId": "allBank",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bank": {
            "put": {
                "tags": [
                    "Bank"
                ],
                "summary": "Edit Bank by id",
                "description": "Edit Bank by id",
                "operationId": "editBank",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Bank",
                        "example": "1"
                    },
                    {
                        "name": "nama_bank",
                        "in": "query",
                        "example": "BCA"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bank"
                ],
                "summary": "Create Bank",
                "description": "Create Bank",
                "operationId": "createBank",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nama_bank": {
                                        "type": "string",
                                        "example": "BCA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bank"
                ],
                "summary": "Delete Bank",
                "description": "Delete Bank by id Bank",
                "operationId": "deleteBank",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id Bank",
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}